# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AddToCartRequest:
    properties:
      book_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - book_id
    - quantity
    type: object
...
  domain.BuyNowRequest:
    properties:
      book_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - book_id
    - quantity
    type: object
  domain.BuyNowResponse:
    properties:
      session_id:
        type: string
    type: object
  domain.CartItem:
    properties:
      book_id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  domain.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.CartItem'
        type: array
      total_price:
        type: number
    type: object
...
  domain.CheckoutRequest:
    properties:
      address_id:
        description: |-
          AddressID is the alias_id UUID of the delivery address; resolved to the internal
          int64 FK before the PostgreSQL transaction.
        type: string
      note:
        type: string
      session_id:
        description: for Buy-Now flow
        type: string
    type: object
...
  domain.Order:
    properties:
      alias_id:
        type: string
      created_at:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      note:
        type: string
      status:
        $ref: '#/definitions/domain.OrderStatus'
      total_amount:
        type: number
    type: object
  domain.OrderItem:
    properties:
      book_id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
  domain.OrderListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/domain.Order'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  domain.OrderStatus:
    enum:
    - pending
    - confirmed
    - packing
    - shipping
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusConfirmed
    - OrderStatusPacking
    - OrderStatusShipping
    - OrderStatusCompleted
    - OrderStatusCancelled
...
  domain.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  server.errorResponse:
    properties:
      error:
        type: string
    type: object
  server.successResponse:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@bookstore.local
    name: API Support
  description: Polyglot persistence bookstore backend — PostgreSQL · MongoDB · Neo4j
    · Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HCMUS Bookstore API
  version: "2.0"
paths:
...
  /cart:
    get:
      description: Return the current user's cart items with enriched metadata
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CartResponse'
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: Add a book to the user's persistent cart in PostgreSQL and sync
        to Redis cache
      parameters:
      - description: Cart item payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Add to cart
      tags:
      - cart
  /cart/{bookId}:
    delete:
      description: Delete a book from the user's cart
      parameters:
      - description: Book MongoDB ID
        in: path
        name: bookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
      security:
      - BearerAuth: []
      summary: Remove cart item
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Update the quantity of a book in the user's cart
      parameters:
      - description: Book MongoDB ID
        in: path
        name: bookId
        required: true
        type: string
      - description: New quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - cart
...
  /orders:
    get:
      description: Return a paginated list of the current user's orders
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OrderListResponse'
      security:
      - BearerAuth: []
      summary: Get order history
      tags:
      - orders
  /orders/{id}:
    get:
      description: Return full order details including line items
      parameters:
      - description: Order Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Get order detail
      tags:
      - orders
  /orders/buy-now:
    post:
      consumes:
      - application/json
      description: Create a temporary 15-minute session for a single-book purchase
      parameters:
      - description: Buy Now payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BuyNowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BuyNowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Buy Now
      tags:
      - orders
  /orders/checkout:
    post:
      consumes:
      - application/json
      description: Place an order from the cart or a Buy-Now session (Atomic Transaction)
      parameters:
      - description: Checkout payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: Checkout
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"